Add Company(Service) & Add Some Converter To SystemMapper Interface.

+++ src/main/java/com/luv2code/demo/service/impl/CategoryService.java
@@ -41,6 +41,10 @@ public class CategoryService implements ICategoryService {
 
 		Optional<Category> category = categoryRepository.findByName(name);
 
+		if (category.isEmpty()) {
+			throw new NotFoundException(NotFoundTypeException.CATEGORY + " Not Found!");
+		}
+		
 		fileHelper.deleteImageFromFileSystem(category.get().getImageUrl());
 
 		categoryRepository.delete(category.get());
